Refactored file names to match widget names and changed the "Files" to "Multifiles"
