+++ orangecontrib/infrared/tests/test_owhyper.py
@@ -7,7 +7,7 @@ from AnyQt.QtCore import QPointF
 import Orange
 from Orange.widgets.tests.base import WidgetTest
 
-from orangecontrib.infrared.widgets.owhyper import values_to_linspace, \
+from orangecontrib.infrared.widgets.owhyperspectra import values_to_linspace, \
     index_values, OWHyper, location_values
 from orangecontrib.infrared.preprocess import Interpolate
 from orangecontrib.infrared.widgets.line_geometry import in_polygon, is_left
