+++ orangecontrib/infrared/tests/test_owmapbuilder.py
@@ -2,7 +2,7 @@ import numpy as np
 import Orange
 from Orange.widgets.tests.base import WidgetTest
 from Orange.widgets.utils.annotated_data import get_next_name
-from orangecontrib.infrared.widgets.owmapbuilder import OWMapBuilder
+from orangecontrib.infrared.widgets.owreshape_map import OWMapBuilder
 
 
 class TestOWFiles(WidgetTest):
