+++ orangecontrib/infrared/widgets/owifg_to_sp.py
