+++ orangecontrib/infrared/widgets/owreshape_map.py
