+++ orangecontrib/infrared/widgets/owspectra.py
